move ota_client_stub.py as grpc.api_v2.servicer
