+++ src/otaclient/app/main.py
@@ -26,7 +26,7 @@ import grpc.aio
 from otaclient import __version__
 from otaclient.app.configs import config as cfg
 from otaclient.app.configs import ecu_info, server_cfg
-from otaclient.app.ota_client_stub import OTAClientServiceStub
+from otaclient.grpc.api_v2.servicer import OTAClientServiceStub
 from otaclient.log_setting import configure_logging
 from otaclient_api.v2 import otaclient_v2_pb2_grpc as v2_grpc
 from otaclient_api.v2.api_stub import OtaClientServiceV2
